Added to dispose method (but untested)
So yeah, I added to the dispose method.  Also, instead of using a List
to keep track of all Characters, I decided to use the MapList class
(which I added in the previous commit) within a Map within a Map.  And I
still haven't done the collision stuff yet... but I will soon.