See description
-Renamed AntiMarioEnemy into GenericEnemy
-Added in an interface called Person in order to relate Player with
NPC's when they are later added in
-Changed DangerZone to extend GenericEnemy (since it is pretty much a
stationary enemy)
-Added a linked list class to be used with enemies so that they can
continually give damage to all the people that they are currently
touching
-GenericEnemy passes through the Player (but will still detect collision
and apply damage).  The reason I did this is so that the Player can
still escape even after being cornered by a swarm of enemies.  If I
didn't do this, each GenericEnemy would just act like a solid wall and
prevent the Player from moving past it